Changed file structuring, moved app.jsx contents
Created new files for the pages that will be used, they are now located in the pages folder. All of the previous app.jsx was moved into the userDashboard.jsx. New App.jsx has a router from react-router-dom to allow navigation between pages.
