+++ frontend/home.html
@@ -1,0 +1,13 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Cat-a-log - Login</title>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="module" src="/src/login.jsx"></script>
+  </body>
+</html>
