+++ frontend/package.json
@@ -13,6 +13,7 @@
     "axios": "^1.7.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.25.1",
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
