+++ frontend/src/components/pages/Login.css
@@ -1,0 +1,3 @@
+.background {
+    background-color: #ffc0cb;
+}
