+++ frontend/src/components/pages/Login.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import './Login.css';
+
+const Login = () => {
+    return(
+            <div className='background'>
+                <h1>Login page</h1>
+            </div>
+    )
+}
+
+export default Login
+
